fix: cache initial data
It appears some time ago the `initialData` configuration was used as a fallback.  When #211 was merged, this behavior changed to be used with SSR like in the next.js example in the README.  Issue #230 explains this was the expectation. I'm using SSR, so im fine with the new behavior.

Since `initialData` is now not quickly revalidated, another issue (#308) has been raised.  Since `initialData` is not cached, and the mutate w/ callback grabs the curerent data from the cache, when `initialData` is used, `undefined` is returned.

fixes #308