Initial implementation of `useSWRList`
Implemented in accordance with the [RFC][1]. Initial implementation is
mostly a copy/paste of `useSWR`, then moved things into arrays so we
can handle the list aspect of this correctly. Plenty of work still to be done,
especially around the duplication between `useSWRList` & `useSWR`,
plus more tests need to be written, but this gives us a basis for discussion.

  [1]: https://github.com/vercel/swr/discussions/1988