Apply suggestions from code review
Hi huozhi,

Thanks for taking the time to review my pull request! I apologize for not noticing that Zeit is now Vercel. I've updated the text to reflect the change, and I hope that the update is now satisfactory.

I appreciate your help and guidance, and I hope that you will consider accepting my pull request.

Co-authored-by: Jiachi Liu <inbox@huozhi.im>